refactor: extract shared BadgeBuilderCore, unify builder across all pages
The badge builder was duplicated between the landing page (badge-builder.tsx)
and the showcase submit dialog (showcase-submit-dialog.tsx) — with the dialog
still using the old 8-provider dropdown UI.

Now:
- lib/badge-builder-shared.ts: shared state type, defaults, presets, URL builder
- components/badge-builder-core.tsx: shared UI (preview, path input, presets,
  controls, customization panel) with children slot for parent-specific content
- components/badge-builder.tsx: landing page wrapper (adds copy output)
- components/showcase-submit-dialog.tsx: showcase wrapper (adds submit flow)

Both use the same builder with identical UX: path input, quick start presets,
variant/size/mode controls, and the customize panel.
